v0.95

+++ src/cryptor.c
@@ -3,15 +3,10 @@
  */
 
 #include "cryptor.h"
-#include <stdlib.h>
 #include <string.h>
-#include <unistd.h>
-#include <openssl/evp.h>
 #include <assert.h>
 
-void
-crypto_md5(const uint8_t *data, size_t data_len, uint8_t *buf)
-{
+void crypto_md5(const uint8_t *data, size_t data_len, uint8_t *buf) {
     EVP_Digest(data, data_len, buf, NULL, EVP_md5(), NULL);
 }
 
@@ -23,11 +18,7 @@ crypto_md5(const uint8_t *data, size_t data_len, uint8_t *buf)
  *   aes-192 的 key 为 24 字节，iv 为 16 字节
  *   aes-256 的 key 为 32 字节，iv 为 16 字节
  */
-void
-bytes_to_key(const uint8_t *passwd,
-             uint8_t *key, size_t key_len,
-             uint8_t *iv, size_t iv_len)
-{
+void bytes_to_key(const uint8_t *passwd, uint8_t *key, size_t key_len, uint8_t *iv, size_t iv_len) {
     uint8_t buf[128];
 
     size_t key_and_iv_len = key_len + iv_len;
@@ -39,8 +30,8 @@ bytes_to_key(const uint8_t *passwd,
     size_t buf_len = MD5_LEN;
 
     while (buf_len < key_and_iv_len) {
-        memcpy(buf+buf_len, passwd, passwd_len);
-        crypto_md5(buf+(buf_len-MD5_LEN), passwd_len+MD5_LEN, buf+buf_len);
+        memcpy(buf + buf_len, passwd, passwd_len);
+        crypto_md5(buf + (buf_len - MD5_LEN), passwd_len + MD5_LEN, buf + buf_len);
         buf_len += MD5_LEN;
     }
 
@@ -51,8 +42,7 @@ bytes_to_key(const uint8_t *passwd,
     }
 }
 
-const EVP_CIPHER *
-get_cipher(const char *cipher_name) {
+const EVP_CIPHER *get_cipher(const char *cipher_name) {
     if (!strncmp(cipher_name, "aes-128-ctr", MAX_CIPHER_NAME_LEN)) {
         return EVP_aes_128_ctr();
     } else if (!strncmp(cipher_name, "aes-256-ctr", MAX_CIPHER_NAME_LEN)) {
@@ -66,16 +56,14 @@ get_cipher(const char *cipher_name) {
     return NULL;
 }
 
-NooneCryptorInfo *
-init_noone_cryptor_info(const char *name,
-        const uint8_t *passwd, size_t key_len, size_t iv_len)
-{
+NooneCryptorInfo *init_noone_cryptor_info(const char *name, const uint8_t *passwd, uint8_t key_len,
+                                          uint8_t iv_len) {
     NooneCryptorInfo *ci = malloc(sizeof(NooneCryptorInfo));
     if (ci == NULL) {
         return NULL;
     }
 
-    size_t name_len = strlen(name);
+    uint8_t name_len = (uint8_t)strlen(name);
     memcpy(ci->cipher_name, name, name_len);
     ci->cipher_name[name_len] = 0;
     ci->cipher_name_len = name_len;
@@ -87,17 +75,13 @@ init_noone_cryptor_info(const char *name,
     return ci;
 }
 
-void
-free_noone_cryptor_info(NooneCryptorInfo *cryptor_info)
-{
-    assert(cryptor_info!=NULL);
+void free_noone_cryptor_info(NooneCryptorInfo *cryptor_info) {
+    assert(cryptor_info != NULL);
 
     free(cryptor_info);
 }
 
-NooneCipherCtx *
-init_noone_cipher_ctx()
-{
+NooneCipherCtx *init_noone_cipher_ctx() {
     NooneCipherCtx *cipher_ctx = malloc(sizeof(NooneCipherCtx));
     if (cipher_ctx == NULL) {
         return NULL;
@@ -108,10 +92,8 @@ init_noone_cipher_ctx()
     return cipher_ctx;
 }
 
-void
-free_noone_cipher_ctx(NooneCipherCtx *cipher_ctx)
-{
-    assert(cipher_ctx!=NULL);
+void free_noone_cipher_ctx(NooneCipherCtx *cipher_ctx) {
+    assert(cipher_ctx != NULL);
 
     if (cipher_ctx->encrypt_ctx != NULL) {
         EVP_CIPHER_CTX_free(cipher_ctx->encrypt_ctx);
@@ -122,10 +104,8 @@ free_noone_cipher_ctx(NooneCipherCtx *cipher_ctx)
     free(cipher_ctx);
 }
 
-EVP_CIPHER_CTX *
-init_evp_cipher_ctx(const EVP_CIPHER *cipher,
-        const uint8_t *key, const uint8_t *iv, int op)
-{
+EVP_CIPHER_CTX *init_evp_cipher_ctx(const EVP_CIPHER *cipher, const uint8_t *key, const uint8_t *iv,
+                                    int op) {
     EVP_CIPHER_CTX *ctx = EVP_CIPHER_CTX_new();
     if (ctx == NULL) {
         return NULL;
@@ -140,10 +120,7 @@ init_evp_cipher_ctx(const EVP_CIPHER *cipher,
 /*
  * 加密失败，返回 0。
  */
-size_t
-encrypt(EVP_CIPHER_CTX *ctx, uint8_t *plaintext,
-        size_t plaintext_len, uint8_t *ciphertext)
-{
+size_t encrypt(EVP_CIPHER_CTX *ctx, uint8_t *plaintext, ssize_t plaintext_len, uint8_t *ciphertext) {
     size_t ciphertext_len;
     int outlen;
 
@@ -152,7 +129,7 @@ encrypt(EVP_CIPHER_CTX *ctx, uint8_t *plaintext,
     }
     ciphertext_len = (size_t)outlen;
 
-    if (!EVP_EncryptFinal_ex(ctx, ciphertext+outlen, &outlen)) {
+    if (!EVP_EncryptFinal_ex(ctx, ciphertext + outlen, &outlen)) {
         return 0;
     }
     ciphertext_len += outlen;
@@ -163,10 +140,8 @@ encrypt(EVP_CIPHER_CTX *ctx, uint8_t *plaintext,
 /*
  * 解密失败，返回 0。
  */
-size_t
-decrypt(EVP_CIPHER_CTX *ctx, uint8_t *ciphertext,
-        size_t ciphertext_len, uint8_t *plaintext)
-{
+size_t decrypt(EVP_CIPHER_CTX *ctx, uint8_t *ciphertext, size_t ciphertext_len,
+               uint8_t *plaintext) {
     size_t plaintext_len;
     int outlen;
 
@@ -175,7 +150,7 @@ decrypt(EVP_CIPHER_CTX *ctx, uint8_t *ciphertext,
     }
     plaintext_len = (size_t)outlen;
 
-    if (!EVP_DecryptFinal_ex(ctx, plaintext+outlen, &outlen)) {
+    if (!EVP_DecryptFinal_ex(ctx, plaintext + outlen, &outlen)) {
         return 0;
     }
     plaintext_len += outlen;
@@ -186,12 +161,10 @@ decrypt(EVP_CIPHER_CTX *ctx, uint8_t *ciphertext,
 /*
  * len 是 4 的 倍数
  */
-void
-rand_bytes(uint8_t *buf, int len)
-{
+void rand_bytes(uint8_t *buf, int len) {
     srand((unsigned int)time(NULL));
-    for (int i = 0; i < len; i+=sizeof(int)) {
+    for (int i = 0; i < len; i += sizeof(int)) {
         int tmp = rand();
-        memcpy(buf+i, &tmp, sizeof(int));
+        memcpy(buf + i, &tmp, sizeof(int));
     }
 }
